requirements.txt
geocoder
pillow
scikit-learn
pandas
numpy
